agents/mcp-server-agent/pyproject.toml
[tool.poetry]
name = "mcp-server-agent"
version = "0.1.0"
description = "MCP server with Nevermined payment-protected tools"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
payments-py = {version = ">=1.1.0", extras = ["mcp"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
openai = "^1.40.0"
python-dotenv = "^1.0.0"

# NOTE: package-mode = false means poetry scripts won't install.
# Use: poetry run python -m src.setup / src.server / src.client

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
